Added missingtypes List
